requirements.txt
Django==2.1.3
djangorestframework==3.8.2
djangorestframework-gis==0.13
GDAL==2.3.2
mapnik==0.1
psycopg2==2.7.6.1
pytz==2018.7
six==1.11.0
